% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_spline.R
\name{expand_spline}
\alias{expand_spline}
\title{Expand feature matris using basis splines}
\usage{
expand_spline(x, df = 3, degree = 3, type = "bs")
}
\arguments{
\item{x}{The design matrix. Columns must represent numeric variables.}

\item{df}{Degrees of freedom; default is 3.}

\item{degree}{Degree of the piecewise polynomial; default is 3 (cubic splines).}

\item{type}{Specifies type of splines; default is \code{"bs"}.}
}
\value{
An object of class \code{expandedMatrix} consisting of:
\itemize{
\item \code{X}: A matrix of dimension \verb{(nrow(x)} by \code{df*ncol(x)}
\item \code{group}: A vector of length \code{df*ncol(x)} that describes the grouping structure
\item Additional metadata on the splines, such as knot locations, required in order to evaluate spline at new feature values (e.g., for prediction)
}
}
\description{
Performs a basis expansion for many variables at once, returning output that is compatible
for use with the \code{grpreg()} function. Returns an expanded matrix along with a vector
that describes its grouping.
}
\details{
\code{expand_spline()} uses the function \code{\link[splines:bs]{splines::bs()}} or \code{\link[splines:ns]{splines::ns()}} to generate a basis
matrix for each column of \code{x}. These matrices represent the spline basis for piecewise
polynomials with specified degree evaluated separately for each original column of \code{x}.
These matrices are then column-bound to form a single grouped design matrix. A vector that
describes the grouping present in the resulting matrix is also generated. The resulting
object can be passed to \code{\link[=grpreg]{grpreg()}}.

This is known as SPAM (INCLUDE WASSERMAN REF)
}
\examples{
X <- expand_spline(attitude[-1])
fit <- grpreg(X, attitude$rating, penalty="grLasso")
plot(fit)
plot_spline(fit, "complaints", lambda=0.001)
}
\seealso{
\code{\link[=plot_spline]{plot_spline()}} to visualize the resulting nonlinear fits
}
