% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_nonlinear_data.R
\name{gen_nonlinear_data}
\alias{gen_nonlinear_data}
\title{Generate example data}
\usage{
gen_nonlinear_data(n = 100, p = 16, seed = 1)
}
\arguments{
\item{n}{Sample size (numeric; default = 100).}

\item{p}{Number of features (numeric; default = 16).}

\item{seed}{Change to get different random data sets (numeric; default = 1).}
}
\description{
Includes nonlinear functions so that basis expansion models can be tested; intended for use with \code{expand_spline()}.
}
\examples{
Data <- gen_nonlinear_data()
}
