% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpmat.R
\name{grpmat}
\alias{grpmat}
\title{Spline Basis for Multiple Variables}
\usage{
grpmat(x, df = 4, degree = 3, type = "bs")
}
\arguments{
\item{x}{the design matrix. Columns must represent numeric variables.}

\item{df}{degrees of freedom; default is 4.}

\item{degree}{degree of the piecewise polynomial; default is 3 for cubic
splines.}

\item{type}{specifies type of splines, default is \code{"bs"}}
}
\value{
A \code{grouped_hat} object composed of a matrix of dimension
\code{c(nrow(x), df*ncol(x))} and a vector of length
\code{df*ncol(x)}
}
\description{
\code{grpmat} Performs a basis expansion for many variables at once for
later use with the grpreg function. Fuction returns one large matrix
and a vector that describes its grouping.
}
\details{
\code{grpmat} is based on the function \code{\link[splines:bs]{splines::bs()}}. It takes each
column of \code{x} and generates a basis matrix. These basis
matrices represent the family of piecewise polynomials with the specified
degree evaluated at the column values of \code{x}. These matrices are then
column-bound to form a single grouped design matrix. A vector that describes
the grouping present in the resulting matrix is also generated. The resulting
object can be passed to \code{\link[=grpreg]{grpreg()}}.
}
\examples{

X <- grpmat(attitude[-1], df = 3)
fit <- grpreg(X, attitude$rating, penalty="grLasso")
plot(fit)

}
