% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdr.R
\name{mfdr}
\alias{mfdr}
\title{Marginal false discovery rates}
\usage{
mfdr(fit, X)
}
\arguments{
\item{fit}{A \code{grpreg} or \code{grpsurv} object.}

\item{X}{The model matrix corresponding to \code{fit}. This is not necessary
for linear regression, but in logistic and Cox regression, the mFDR depends
on X. It is not necessary to supply \code{X} if it is already contained in
\code{fit}; i.e., if \code{ncvreg}/\code{ncvsurv} was run with \code{returnX = TRUE}.}
}
\value{
An object with S3 class \code{mfdr} inheriting from \code{data.frame}, containing:
\describe{
\item{ef}{The number of variables selected at each value of \code{lambda},
averaged over the permutation fits.}
\item{s}{The actual number of selected variables for the non-permuted data.}
\item{mfdr}{The estimated marginal false discovery rate (\code{ef/s}).}
}
}
\description{
Estimates the marginal false discovery rate (mFDR) of a group penalized
regression model.
}
\details{
The function estimates the marginal false discovery rate (mFDR) for groups in
a group lasso or group MCP penalized regression model. The estimate tends to
be accurate in most settings, but will be somewhat conservative if predictors
are highly correlated.
}
\examples{
# Birthweight data ---------------------------
data(Birthwt)
x <- Birthwt$X
group <- Birthwt$group

# Linear regression --------------------------
y <- Birthwt$bwt
fit <- grpreg(x, y, group)
obj <- mfdr(fit)
head(obj)

# Logistic regression ------------------------------
y <- Birthwt$low
fit <- grpreg(x, y, group, penalty="grMCP", family="binomial", returnX = TRUE)
obj <- mfdr(fit)
# If returnX is not TRUE, user must supply X
fit <- grpreg(x, y, group, penalty="grMCP", family="binomial")
obj <- mfdr(fit, x)
head(obj)

# Cox regression -----------------------------------
data(Lung)
x_lung <- Lung$X
y_lung <- Lung$y
g_lung <- Lung$group
fit <- grpsurv(x_lung, y_lung, g_lung, penalty = "grSCAD")
obj <- mfdr(fit, x_lung)
head(obj)
}
\seealso{
\code{\link[=grpreg]{grpreg()}}, \code{\link[=grpsurv]{grpsurv()}}
}
