% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sp.R
\name{plot.sp.grpreg}
\alias{plot.sp.grpreg}
\title{Plot Spline Curves}
\usage{
\method{plot}{sp.grpreg}(
  fit,
  variable,
  lambda,
  which = NULL,
  scatter = FALSE,
  type = "contrast",
  warnings = TRUE,
  ...
)
}
\arguments{
\item{fit}{A \code{grpreg} object. The model must have been fit using a
\code{grpmat} object.}

\item{variable}{The name of the variable which will be plotted.}

\item{lambda}{Values of the regularization parameter \code{lambda}
which will be used for the plot. Each value of lambda will produce
a different curve.}

\item{which}{Indices of the penalty parameter \code{lambda} which
will be used for the plot. If \code{lambda} is specified, this will
override \code{which}.}

\item{scatter}{If \code{TRUE}, a scatter plot of the partial residuals will
also be produced. Default is \code{FALSE}. If multiple lambdas are
specified, the largest value will be used to calculate the residuals.}

\item{type}{Type of plot to be produced. Default is \code{contrast}.
The following options are supported:
\itemize{
\item If \code{conditional} is selected, the plot returned shows the value of
the variable on the x-axis and the change in response on the y-axis,
holding all other variables constant at their mean value.
\item If \code{contrast} is selected, the plot returned shows the effect
on the expected value of the response by moving the x variable away
from the mean on the x-axis.
}}

\item{warnings}{If \code{FALSE}, warnings will be supressed. Default is
\code{TRUE}.}
}
\description{
\code{plot.sp} Plots a spline curve for a single variable using a
grpreg object.
}
\details{
\code{plot.sp} takes a model fit using both the \code{\link[=grpreg]{grpreg()}} and \code{\link[=grpmat]{grpmat()}}
functions and plots a spline curve for a given variable.
}
\examples{

X <- grpmat(attitude[-1], df = 3)
fit <- grpreg(X, attitude$rating, penalty="grLasso")
plot.sp(fit, "complaints", which = c(45, 50))


}
