% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spline.R
\name{plot_spline}
\alias{plot_spline}
\title{Plot spline curve}
\usage{
plot_spline(
  fit,
  variable,
  lambda,
  which = NULL,
  scatter = FALSE,
  type = "contrast",
  warnings = TRUE,
  ...
)
}
\arguments{
\item{fit}{A \code{grpreg} object. The model must have been fit using a \code{expand_spline} object.}

\item{variable}{The name of the variable which will be plotted.}

\item{lambda}{Values of the regularization parameter \code{lambda} which will be used for the plot. Each value of lambda will produce a different curve.}

\item{which}{Indices of the penalty parameter \code{lambda} which will be used for the plot. If both \code{lambda} and \code{which} are specified, \code{lambda} takes precedence.}

\item{scatter}{If \code{TRUE}, a scatter plot of the partial residuals will also be produced. Default is \code{FALSE}. If multiple lambdas are specified, the largest value will be used to calculate the residuals.}

\item{type}{Type of plot to be produced. Default is \code{contrast}.  The following options are supported:
* If \code{conditional} is selected, the plot returned shows the value of the variable on the x-axis and the change in response on the y-axis, holding all other variables constant at their mean value.
* If \code{contrast} is selected, the plot returned shows the effect on the expected value of the response by moving the x variable away from the mean on the x-axis.}

\item{warnings}{If \code{FALSE}, warnings will be suppressed. Default is \code{TRUE}.}

\item{...}{Further arguments to be passed to \code{plot()}}
}
\description{
Plots a spline curve for a single variable using a grpreg object for which an additive model was fit.
}
\details{
\code{plot_spline()} takes a model fit using both the \code{\link[=grpreg]{grpreg()}} and \code{\link[=expand_spline]{expand_spline()}} functions and plots a spline curve for a given variable.
}
\examples{
X <- expand_spline(attitude[-1], df = 3)
fit <- grpreg(X, attitude$rating, penalty="grLasso")
plot_spline(fit, "complaints", which = c(5, 90))
}
